refactor: share thread binding id parser
(cherry picked from commit 30d091b2fb2e37f7fe967b78403dbdcb1287a350)
