+++ src/telegram/thread-bindings.ts
@@ -1,6 +1,7 @@
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
+import { resolveThreadBindingConversationIdFromBindingId } from "../channels/thread-binding-id.js";
 import { resolveStateDir } from "../config/paths.js";
 import { formatThreadBindingDurationLabel } from "../discord/monitor/thread-bindings.messages.js";
 import { logVerbose } from "../globals.js";
@@ -312,22 +313,6 @@ async function persistBindingsToDisk(params: {
   });
 }
 
-function resolveThreadIdFromBindingId(params: {
-  accountId: string;
-  bindingId?: string;
-}): string | undefined {
-  const bindingId = params.bindingId?.trim();
-  if (!bindingId) {
-    return undefined;
-  }
-  const prefix = `${params.accountId}:`;
-  if (!bindingId.startsWith(prefix)) {
-    return undefined;
-  }
-  const conversationId = bindingId.slice(prefix.length).trim();
-  return conversationId || undefined;
-}
-
 function normalizeTimestampMs(raw: unknown): number {
   if (typeof raw !== "number" || !Number.isFinite(raw)) {
     return Date.now();
@@ -569,7 +554,7 @@ export function createTelegramThreadBindingManager(
         : null;
     },
     touch: (bindingId, at) => {
-      const conversationId = resolveThreadIdFromBindingId({
+      const conversationId = resolveThreadBindingConversationIdFromBindingId({
         accountId,
         bindingId,
       });
@@ -592,7 +577,7 @@ export function createTelegramThreadBindingManager(
           }),
         );
       }
-      const conversationId = resolveThreadIdFromBindingId({
+      const conversationId = resolveThreadBindingConversationIdFromBindingId({
         accountId,
         bindingId: input.bindingId,
       });
